fix spork int error

+++ src/spork.cpp
@@ -103,7 +103,7 @@ bool IsSporkActive(int nSporkID)
 // grab the value of the spork on the network, or the default
 int GetSporkValue(int nSporkID)
 {
-    int r = 0;
+    int64_t r = 0;
 
     if(mapSporksActive.count(nSporkID)){
         r = mapSporksActive[nSporkID].nValue;
